feat: 新增权重完整继承功能并清理冗余文件
- 新增权重完整继承API，支持专家打分→基准表(区县→市→省)的继承链
- 使用modelId替代configName作为主要查找参数，提高可靠性
- 评估历史页面新增"评估名称"列，显示年份+区县+模型信息
- 添加SQL脚本目录，包含表结构和数据初始化脚本
- 清理冗余配置文件和测试工具类(.cursor, .kiro, .superdesign, .trae, nginx等)

+++ src/main/java/com/evaluate/entity/SurveyData.java
@@ -6,6 +6,7 @@ import lombok.EqualsAndHashCode;
 import lombok.experimental.Accessors;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
@@ -225,7 +226,7 @@ public class SurveyData implements Serializable {
      * 上一年度防灾减灾救灾资金投入总金额(万元)
      */
     @TableField("funding_amount")
-    private Double fundingAmount;
+    private BigDecimal fundingAmount;
 
     /**
      * 救灾物资储备方式
@@ -285,7 +286,7 @@ public class SurveyData implements Serializable {
      * 现有储备物资、装备折合金额(万元)
      */
     @TableField("material_value")
-    private Double materialValue;
+    private BigDecimal materialValue;
 
     /**
      * 本级灾害应急避难场所数量(个或处)
